ran migrations

+++ db/migrate/20160418181402_create_generic_files.rb
@@ -3,7 +3,7 @@ class CreateGenericFiles < ActiveRecord::Migration
     create_table :generic_files do |t|
       t.string :file_format
       t.string :uri
-      t.long :size
+      t.float :size
       t.string :identifier
       t.string :intellectual_object
       t.belongs_to :intellectual_object, index: true
